Make looping coroutines backed by a notifier
* launchFrequency moved into wpi subproject, as it now depends on the
pressence of a notifier
* Rather than using delay(), which is backed by Thread.sleep, change
coroutine looping function to use a notifier. Notifiers are much more
precise for timing. In testing, the standard deviation of delta times
for notifier-back loops is 4 orders of magnitude smaller than that of
Thread.sleep backed loops
* Add FalconNotifier, which acts similar to Notifier, except lower
level. Provides a minimal wrapper over HAL calls and allows the user to
use the notifier as they see fit